Add files via upload
This is my R codes for implementing the non-self-referential Gram-Schimdt Process in the manuscript. In the code, please look directly into the main program NsrGSP function with two inputs: the input vector U (regarded as a set of column vectors) and kk (the number of intended orthogonal vectors to be generated. Be warned, kk should be less than the number of column vectors in U.  In addition, if one wants to use this function, you could simply call it from your code by "source(file.choose())".